Moved Face Detection into its own component
