+++ client/src/components/FileUpload.jsx
@@ -1,19 +1,22 @@
 import React, { createRef } from 'react';
 
 function FileUpload(props) {
+
+  const { setImageFile, setImageURL } = props;
+
   let uploadInput = createRef();
 
   const handleUploadImage = event => {
     event.preventDefault();
     const data = new FormData();
-    props.setImageFile(uploadInput.files[0]);
+    setImageFile(uploadInput.files[0]);
     data.append('file', uploadInput.files[0]);
     fetch('http://localhost:3001/upload', {
       method: 'POST',
       body: data,
     }).then((response) => {
       response.json().then((body) => {
-        props.setImageURL(`http://localhost:3001/${body.file}`);
+        setImageURL(`http://localhost:3001/${body.file}`);
       });
     });
   }
